requirements.txt
openai==1.42.0
PyGithub==2.3.0
requests==2.31.0
